[oclif-plugin] Break out utils.ts
