+++ packages/plugins/typescript/oclif/src/visitor.ts
@@ -1,9 +1,9 @@
-import { ClientSideBaseVisitor, ClientSideBasePluginConfig, LoadedFragment, getConfigValue, OMIT_TYPE, indentMultiline, RawClientSideBasePluginConfig } from '@graphql-codegen/visitor-plugin-common';
+import { ClientSideBaseVisitor, ClientSideBasePluginConfig, LoadedFragment, indentMultiline } from '@graphql-codegen/visitor-plugin-common';
 import * as autoBind from 'auto-bind';
-import { GraphQLNamedType, GraphQLSchema, NamedTypeNode, OperationDefinitionNode, print, TypeNode, VariableDefinitionNode } from 'graphql';
-import { join, resolve } from 'path';
+import { GraphQLSchema, OperationDefinitionNode, print } from 'graphql';
 
 import { Config, Info } from '.';
+import { getFlagConfigForVariableDefinition, omitOclifDirectives } from './utils';
 
 interface Operation {
   node: OperationDefinitionNode;
@@ -18,8 +18,6 @@ interface OclifDirectiveValues {
   examples?: string[];
 }
 
-const oclifDirectiveKeys: Array<keyof OclifDirectiveValues> = ['description', 'examples'];
-
 export class GraphQLRequestVisitor extends ClientSideBaseVisitor<Config, ClientSideBasePluginConfig> {
   private _operationsToInclude: Operation[] = [];
   private _info: Info;
@@ -29,9 +27,10 @@ export class GraphQLRequestVisitor extends ClientSideBaseVisitor<Config, ClientS
     this._info = info;
 
     const { clientPath = '../../client' } = rawConfig;
+
     // FIXME: This is taken in part from
     //  presets/near-operation-file/src/index.ts:139. How do I build a path relative to the outputFile in the same way?
-    //  A plugin doesn't appear to have access to the same "options.baseOutputDir" that this preset does.
+    //  A plugin doesn't appear to have access to the same "options.baseOutputDir" that the preset does.
     // const absClientPath = resolve(info.outputFile, join(options.baseOutputDir, options.presetConfig.baseTypesPath));
 
     autoBind(this);
@@ -65,7 +64,6 @@ export class GraphQLRequestVisitor extends ClientSideBaseVisitor<Config, ClientS
     }
 
     const operation = this._operationsToInclude[0];
-    const resultType = this._schema.getType(operation.operationResultType);
 
     // Find the @oclif directive in the client document, if it's there
     const directive = operation.node.directives.find(directive => directive.name.value === 'oclif');
@@ -119,65 +117,3 @@ ${indentMultiline(flags.join(',\n'), 2)}
 `;
   }
 }
-
-type OclifFlagType = 'boolean' | 'integer' | 'string' | 'enum';
-
-// TODO: Implement enum
-const mapVariableTypeToOclifType = (type: NamedTypeNode): OclifFlagType => {
-  if (type.name.value === 'Boolean') {
-    return 'boolean';
-  } else if (['Float', 'Int'].includes(type.name.value)) {
-    // A quirk of oclif is that "integer" allows for a `number`-typed response, and then
-    //   we supply our own parsing function to make sure it's a float and not an integer
-    return 'integer';
-  } else {
-    return 'string';
-  }
-};
-
-const getParserForType = (type: NamedTypeNode): string | void => {
-  if (type.name.value === 'Float') {
-    return 'input => Number(input)';
-  }
-};
-
-const getFlagConfigForVariableDefinition = (definition: VariableDefinitionNode) => {
-  const { list, required, innerType } = getInnerType(definition.type);
-  const oclifType = mapVariableTypeToOclifType(innerType);
-  const parser = getParserForType(innerType);
-
-  return `${definition.variable.name.value}: flags.${oclifType}({
-  multiple: ${list},
-  required: ${required},${parser ? `\n  parse: ${parser}` : ''}
-})`;
-};
-
-interface TypeConfig {
-  required: boolean;
-  list: boolean;
-  innerType: NamedTypeNode;
-}
-
-const getInnerType = (type: TypeNode): TypeConfig => {
-  let result: Partial<TypeConfig> = {
-    list: false,
-    required: false,
-  };
-  let _type = type;
-  while (_type.kind !== 'NamedType') {
-    if (_type.kind === 'ListType') {
-      result.list = true;
-    } else if (_type.kind === 'NonNullType') {
-      result.required = true;
-    }
-    _type = _type.type;
-  }
-  result.innerType = _type;
-  return result as TypeConfig;
-};
-
-// remove all @oclif directives from the document for transmission to the server
-const omitOclifDirectives = (node: OperationDefinitionNode) => {
-  const directives = node.directives.filter(directive => directive.name.value !== 'oclif');
-  return Object.assign({}, node, { directives });
-};
